edit components App, MovieCard; edit constants, utils; refactoring code; bugs fix
